requirements.txt
asgiref==3.3.4
attrs==21.4.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-allauth==0.49.0
django-bootstrap-modal-forms==2.2.0
django-ckeditor==6.5.0
django-countries==7.2.1
django-crispy-forms==1.13.0
django-environ==0.8.1
django-js-asset==2.0.0
django-widget-tweaks==1.4.12
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
Pillow==8.1.2
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
urllib3==1.26.13
whitenoise==5.3.0
